Move NewText to ui
