+++ libraries/ui/src/default-config/tailwind.css
@@ -72,6 +72,25 @@
     ::file-selector-button {
         border-color: var(--color-stone-200, currentColor);
     }
+
+    .bluedot-h1 {
+        @apply text-color-secondary-text font-sans text-size-xl font-bold;
+    }
+    .bluedot-h2 {
+        @apply text-color-secondary-text font-sans text-size-xl font-bold;
+    }
+    .bluedot-h3 {
+        @apply text-color-secondary-text font-sans text-size-lg font-[650];
+    }
+    .bluedot-h4 {
+        @apply text-color-secondary-text font-sans text-size-md font-[650];
+    }
+    .bluedot-p {
+        @apply text-color-text font-sans text-size-sm font-normal;
+    }
+    .bluedot-a {
+        @apply text-color-secondary-text hover:text-color-primary-accent cursor-pointer underline;
+    }
 }
 
 @utility bluedot-base {
